Added methods to set the Visitor Id in the Tracker class
In order to be able to connect events from a client (Webclient, Android,
iOS) with events created by a backend component (based on some
interactions from the client with the backend), it is very helpful
if the Visitor ID can be set to a specific value.
Therefor methods were added to the Tracker class with which the Visitor
ID can be set and also verified that the Visitor ID format is exactly 16
bytes long and only containing lowercase hexadecimal characters.