Added 'link' key to the Query Parameter Constants
In order to create consistent events from different clients (Webclient,
Android, iOS) it can be useful to set the 'link' value.
Therefor I added a new Query key which can be used with the public
'set' method of the tracker in order to provide a 'link' Uri along with
the event.